fix_coding
